+++ src/Blego/task.js
@@ -14,17 +14,17 @@ const EMOJI = {
  * @instance
  * @memberof Blego
  * @param {string} label Label of the task.
- * @param {function} func Task function.
+ * @param {function} fn Task function.
  */
-function task(label, func) {
+function task(label, fn) {
   this.tools.tryCatch(
     () => {
       this.tools.validateType('label', 'string', label);
-      this.tools.validateType('func', 'function', func);
+      this.tools.validateType('fn', 'function', fn);
 
       console.log(EMOJI.wait, 'Running', quote(chalk.cyan(label)), '...');
       const startTime = process.hrtime();
-      func();
+      fn();
       const diffTime = process.hrtime(startTime);
       const runTime = diffTime[0] + diffTime[1] / 1e9;
       console.log(EMOJI.finish, 'Finished', quote(chalk.cyan(label)), 'in', chalk.magenta(runTime.toFixed(3) + 's'));