+++ src/Blego/tasks/loadGlobals.js
@@ -1,0 +1,16 @@
+const parseDataDir = require('../../tools/parseDataDir.js');
+
+/**
+ * Parses the globals files.
+ *
+ * @private
+ * @instance
+ * @memberof Blego
+ */
+function loadGlobals() {
+  this.task('Load globals', () => {
+    this.global = parseDataDir(this.options.paths.globals).cast();
+  });
+}
+
+module.exports = loadGlobals;