+++ tests/blego.tasks.loadData.test.js
@@ -1,0 +1,30 @@
+describe('blego.tasks.loadData', () => {
+  const mockFs = require('mock-fs');
+  const Blego = require('Blego.js');
+  let blego;
+
+  beforeEach(() => {
+    console.log = jest.fn();
+    blego = new Blego();
+    new blego.Store([]);
+    mockFs({
+      'data/authors/matt.json': '{"name": "matt"}',
+      'data/posts/a.json': '{"id": "a"}',
+      'data/posts/b.json': '{"id": "b"}',
+    });
+  });
+
+  afterEach(() => {
+    mockFs.restore();
+  });
+
+  it('Loads data into stores', () => {
+    blego.tasks.loadData();
+
+    expect(blego.store.authors.count()).toEqual(1);
+    expect(blego.store.posts.count()).toEqual(2);
+    expect(blego.store.authors.get('matt').name).toEqual('matt');
+    expect(blego.store.posts.get('a').id).toEqual('a');
+    expect(blego.store.posts.get('b').id).toEqual('b');
+  });
+});